+++ sample/07.GraphQL/README.md
@@ -44,6 +44,13 @@ go get github.com/99designs/gqlgen/internal/imports@v0.17.9
 go run github.com/99designs/gqlgen generate
 ```
 
+or
+
+```s
+go generate ./...
+```
+
+
 
 ## Update Resolver
